+++ source/fortran/input.f90
@@ -666,7 +666,7 @@ contains
           write(outfile,*) ' '
           obsolete_var = .true.
        case ('ilhthresh')
-          call parse_int_variable('ilhthresh', ilhthresh, 1, 17, &
+          call parse_int_variable('ilhthresh', ilhthresh, 1, 18, &
                'Switch for L-H power threshold to enforce')
        case ('impc')
           call parse_real_variable('impc', impc, 0.0D0, 10.0D0, &