+++ source/fortran/numerics.f90
@@ -104,7 +104,7 @@ module numerics
   !+ad_vars  ipeqns  FIX : number of constraint equations available
   integer, parameter :: ipeqns = 79
   !+ad_vars  ipnfoms FIX : number of available figures of merit
-  integer, parameter :: ipnfoms = 18
+  integer, parameter :: ipnfoms = 19
 
   integer, parameter :: ipvlam  = ipeqns+2*ipnvars+1
   integer, parameter :: iptnt   = (ipeqns*(3*ipeqns+13))/2
@@ -158,8 +158,11 @@ module numerics
        'min R0, max tau_burn. ', &
        !+ad_varc  <LI> (17) net electrical output
        'net electrical output.', &
-       !+ad_varc  <LI> (18) Null Figure of Merit </UL>
-       'Null figure of merit. '  &
+       !+ad_varc  <LI> (18) Null Figure of Merit
+       'Null figure of merit. ',  &
+       !+ad_varc  <LI> (19) linear combination of big Q and pulse length (maximised)
+       !+ad_varc              note: FoM should be minimised only!</UL>
+       'max Q, max t_burn.    ' &
         /)
   !+ad_vars  ncalls : number of function calls during solution
   integer :: ncalls = 0