+++ source/fortran/physics_functions.f90
@@ -181,7 +181,7 @@ contains
     pthrmw(17) = 2.11 * (plascur/1.0D6)**1.18 * dnla20**0.83
 
     ! Hubbard et al. 2017 L-I threshold scaling
-    pthrmw(18) = 0.2 * dnla20 * sarea * (bt/2.0)**0.25
+    pthrmw(18) = 0.162 * dnla20 * sarea * (bt)**0.26
 
   end subroutine pthresh